\name{Table10.7}
\alias{Table10.7}
\docType{data}
\title{
USA Consumption Expenditure
}
\description{
USA Consumption Expenditure from 1947 to 2000
}
\usage{data("Table10.7")}
\format{
  A data frame with 54 observations on the following 5 variables.
  \describe{
    \item{\code{Year}}{Calendar Year}
    \item{\code{C}}{Real Consumption Expenditure in billions of chained 1996 USD}
    \item{\code{Yd}}{Real Disposable Personal Income in billions of chained 1996 USD}
    \item{\code{W}}{Real Wealth in billions of chained 1996 USD}
    \item{\code{I}}{Nominal annual yield on 3-month treasury securities - inflation rate (measured by annual \% change in annual chained price index)}
  }
}
\details{
This table is same as the Table7.12
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Table10.7)
## maybe str(Table10.7) ; plot(Table10.7) ...
}
\keyword{datasets}
