\name{Table11.7}
\alias{Table11.7}
\docType{data}
\title{
Passenger Car Mileage Data
}
\description{
Passenger car mileage data in the USA
}
\usage{data("Table11.7")}
\format{
  A data frame with 81 observations on the following 6 variables.
  \describe{
    \item{\code{HP}}{Engine Horsepower}
    \item{\code{MPG}}{Miles per Gallon}
    \item{\code{MPGF}}{Modified MPG (Not used in the book example)}
    \item{\code{VOL}}{Cubic feet of cab space}
    \item{\code{WT}}{Vehicle weight, 100s of pounds}
    \item{\code{SP}}{Top speed, miles per hour}
  }
}
\details{
Out of these variables, MPGF is not used in the book.
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Table11.7)
## maybe str(Table11.7) ; plot(Table11.7) ...
}
\keyword{datasets}
