\name{Table16.16}
\alias{Table16.16}
\docType{data}
\title{
Unemployment Rate and Hourly Compensation
}
\description{
Unemployment Rate and Hourly Compensation in Manufacturing in Different Countries
}
\usage{data("Table16.16")}
\format{
  A data frame with 27 observations on the following 7 variables.
  \describe{
    \item{\code{Year}}{a numeric vector}
    \item{\code{COMP_US}}{Hourly compensation in the USA}
    \item{\code{COMP_CAN}}{Hourly compensation in Canada}
    \item{\code{COMP_UK}}{Hourly compensation in the UK}
    \item{\code{UN_US}}{Unemployment rate (\%) in the USA}
    \item{\code{UN_CAN}}{Unemployment rate (\%) in Canada}
    \item{\code{UN_UK}}{Unemployment rate (\%) in the UK}
  }
}
\details{
Hourly compensation is measured as index of hourly compensation in US Dollars, 1992 = 100
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Table16.16)
## maybe str(Table16.16) ; plot(Table16.16) ...
}
\keyword{datasets}
