\name{Table16.17}
\alias{Table16.17}
\docType{data}
\title{
Gasoline Demand for 18 Countries
}
\description{
Gasoline Demand for 18 OECD Countries
}
\usage{data("Table16.17")}
\format{
  A data frame with 342 observations on the following 7 variables.
  \describe{
    \item{\code{COUNTRY}}{Country}
    \item{\code{Ctry_Code}}{Country Code}
    \item{\code{YEAR}}{Year of observation}
    \item{\code{LGASPCAR}}{Log of gasoline consumption per car}
    \item{\code{LINCOMEP}}{Log of per capita income}
    \item{\code{LRPMG}}{Log of real price of gasoline}
    \item{\code{LCARPCAP}}{Log of number of car per capita}
  }
}
\details{
Data for Exercise 16.15
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Table16.17)
## maybe str(Table16.17) ; plot(Table16.17) ...
}
\keyword{datasets}
