\name{Table16.18}
\alias{Table16.18}
\docType{data}
\title{
The Determinants of Aid in the Post-Cold War Era
}
\description{
The Determinants of Aid in the Post-Cold War Era
}
\usage{data("Table16.18")}
\format{
  A data frame with 395 observations on the following 20 variables.
  \describe{
    \item{\code{aidcap}}{Per capita aid}
    \item{\code{raidcap}}{real aid per capita (in millions 2000$)}
    \item{\code{infmort}}{infant mortality rate}
    \item{\code{infmort_2}}{Square in infant mortality rate}
    \item{\code{pop}}{Population}
    \item{\code{raid}}{real aid (in millions 2000$)}
    \item{\code{civlib}}{Civil rights index (1 = most restrictive, 7 = most free)}
    \item{\code{polright}}{Political rights index (1 = most restrictive, 7 = most free)}
    \item{\samp{civil/PolRts}}{Sum of civil and political rights}
    \item{\code{goveff}}{government effectiveness}
    \item{\code{regqual}}{a numeric vector}
    \item{\code{rgdpcap1}}{a numeric vector}
    \item{\code{rgdpcap}}{a numeric vector}
    \item{\code{rgdcap_2}}{a numeric vector}
    \item{\code{popmil}}{Population in Millions}
    \item{\code{popmil_2}}{Square of population in Millions}
    \item{\code{deflator}}{a numeric vector}
    \item{\code{y1}}{a numeric vector}
    \item{\code{y2}}{Dummy variable for year 2000}
    \item{\code{y3}}{Dummy variable for year 2003}
  }
}
\details{
Some variables are not clear. Please read the actual paper from the website https://files.stlouisfed.org/files/htdocs/publications/review/07/11/Bandyopadhyay.pdf
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Table16.18)
## maybe str(Table16.18) ; plot(Table16.18) ...
}
\keyword{datasets}
