\name{Table17.7}
\alias{Table17.7}
\docType{data}
\title{
Private Consumption Expenditure and GDP
}
\description{
Private Consumption Expenditure and GDP data from Sri Lanka
}
\usage{data("Table17.7")}
\format{
  A data frame with 27 observations on the following 3 variables.
  \describe{
    \item{\code{YEAR}}{Calendar Year}
    \item{\code{PCON}}{Private consumption expenditure}
    \item{\code{GDP}}{GDP}
  }
}
\details{
Data compiled from the World Bank's World Tables
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Table17.7)
## maybe str(Table17.7) ; plot(Table17.7) ...
}
\keyword{datasets}
