\name{Table18.3}
\alias{Table18.3}
\docType{data}
\title{
Demand and Supply for Gas
}
\description{
Demand and Supply for Gas in the USA
}
\usage{data("Table18.3")}
\format{
  A data frame with 296 observations on the following 7 variables.
  \describe{
    \item{\code{Year}}{Calendar Year}
    \item{\code{Qtr}}{Quarter of the Year}
    \item{\code{OBS}}{Observation}
    \item{\code{CARSALES}}{Millions of cars sold per year}
    \item{\code{PERSINCOME}}{Personal income, $ billions}
    \item{\code{PRICEGAS}}{Price of gas (cents per gallon)}
    \item{\code{QUANTGAS}}{Gas consumption (thousands of barrels per day, unleaded)}
  }
}
\details{
Exercise 18.13 in the book
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Table18.3)
## maybe str(Table18.3) ; plot(Table18.3) ...
}
\keyword{datasets}
