\name{Table20.5}
\alias{Table20.5}
\docType{data}
\title{
Underlying Data for Klein’s Model
}
\description{
Underlying Data for Klein’s Model discussed in Example 18.6 of the textbook
}
\usage{data("Table20.5")}
\format{
  A data frame with 22 observations on the following 10 variables.
  \describe{
    \item{\code{YEAR}}{Calendar Year}
    \item{\samp{C*}}{a numeric vector}
    \item{\code{P}}{a numeric vector}
    \item{\code{W}}{a numeric vector}
    \item{\code{I}}{a numeric vector}
    \item{\samp{K-1}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
    \item{\samp{W'}}{a numeric vector}
    \item{\code{G}}{a numeric vector}
    \item{\code{T}}{a numeric vector}
  }
}
\details{
The data dictionary not available on the textbook's website
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Table20.5)
## maybe str(Table20.5) ; plot(Table20.5) ...
}
\keyword{datasets}
