\name{Table22.5}
\alias{Table22.5}
\docType{data}
\title{
HWI and UN
}
\description{
Help-Wanted Index and Unemployment Rate
}
\usage{data("Table22.5")}
\format{
  A data frame with 373 observations on the following 3 variables.
  \describe{
    \item{\code{TIME}}{1  = January 1969, 373 = January 2000}
    \item{\code{HWI}}{Help-Wanted Index}
    \item{\code{UNRATE}}{ Unemployment rate (\%)}
  }
}
\details{
Data for Example 22.3
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Table22.5)
## maybe str(Table22.5) ; plot(Table22.5) ...
}
\keyword{datasets}
