\name{Table22.6}
\alias{Table22.6}
\docType{data}
\title{
Yen Exchange Rate
}
\description{
Yen Exchange Rate against the USD
}
\usage{data("Table22.6")}
\format{
  A data frame with 448 observations on the following 4 variables.
  \describe{
    \item{\code{exchange}}{Exchange Rate (Yen per USD)}
    \item{\code{Time}}{Months since December 1970}
    \item{\code{Year}}{Calendar Year}
    \item{\code{Month}}{Month of the Year}
  }
}
\details{
Data for Example 22.4
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Table22.6)
## maybe str(Table22.6) ; plot(Table22.6) ...
}
\keyword{datasets}
