\name{Table3.6}
\alias{Table3.6}
\docType{data}
\title{
Productivity and Related Data for Business Sector (1960 - 2005)
}
\description{
Quarterly Productivity and Related Data for Business Sector (1960 - 2005),  with 1992 = 100
}
\usage{data("Table3.6")}
\format{
  A data frame with 46 observations on the following 5 variables.
  \describe{
    \item{\code{Year}}{a numeric vector}
    \item{\code{Output_per_hour_business_sector}}{Real GDP of business sector measured in per hour term}
    \item{\code{Output_per_hour_nonfarm_sector}}{Real GDP of non-farm business sector measured in per hour term}
    \item{\code{Compensation_per_hour_business_sector}}{Hourly compensation for business sector employees}
    \item{\code{Compensation_per_hour_nonfarm_sector}}{Hourly compensation for non-farm sector workers}
  }
}
\details{
1. Compensation include wages and salaries, plus employers' contribution to social security.
2. Hourly compensation is divided by consumer price index for the most recent quarter
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Table3.6)
## maybe str(Table3.6) ; plot(Table3.6) ...
}
\keyword{datasets}
