\name{Table5.10}
\alias{Table5.10}
\docType{data}
\title{
CPI and WPI in India
}
\description{
Consumer Price Index and Wholesale Price Index in India
}
\usage{data("Table5.10")}
\format{
  A data frame with 35 observations on the following 4 variables.
  \describe{
    \item{\code{Year}}{Financial Year}
    \item{\code{CPI_AL}}{CPI for Agriculture Laborers}
    \item{\code{CPI_IW}}{CPI for Industrial Workers}
    \item{\code{CPI_UNE}}{CPI for Non-manual Employees}
    \item{\code{WPI}}{Wholesale Price Index}
  }
}
\details{
For all indices, 1960-61 = 100
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Table5.10)
## maybe str(Table5.10) ; plot(Table5.10) ...
}
\keyword{datasets}
